Added initial support for interrupts with the functions: - bool cxl_pending_events (struct cxl_afu_h *afu); - int cxl_read_event (struct cxl_afu_h *afu, struct cxl_event *event);
THIS CHANGE ADDS A DEPENDENCY ON LATEST 3.18 KERNEL HEADERS!
The user will need /usr/include/misc/cxl.h on their system where they compile
the libcxl dependent code.  Only the headers are need, not the actual kernel.
It is only used to reference structs in the header and not functions.
This file can not be included in this repository as it is covered by a
different open source license which can't be mixed with this code.