Restructure buffer handling so responses can come back out of order.
This commit restructures and cleans up some of the code a bit. I've removed
the first_br/last_br link list and replaced it with an array that is used for
both reads and writes.

Whenever a read/write comes in it is placed into a random slot in the queue. Every
cycle there is a random 1 in 8 chance that one of these queue items will be handled.
Every time we handle an item we choose one at random then either queue the buffer
read or send the write and the completion. If a buffer has been read a completion
may also be sent.

This structure is also a bit nicer as it allows the checks for page, flushing
and bad memory address to be in a common, easy to follow, location.