Abstract away inline assembly in iniline functions.
The main purpose of this commit is so that I can ensure this library builds
in our environment without access to main hardware.

I've done this by implementing the inline assembly in a portable way, fully
understanding that this library won't likely ever be used on another platform.

Also, I wouldn't be surprised if the portable C store/load functions outputed
the same assembly as the inline assembly versions. However, I have not
checked this and I didn't want to make changes, that could have subtle side
effects, to code that works and has been tested.

I also feel this commit improves code readibility a bit.