Modified cxl.h search method.
First we use CC to check if the header exists. Thus if it's already includeable
the user doesn't have to do anything.

The one iffy thing here is if the file is includeable with <misc/cxl.h> then
the user will have to set the HAVE_MISC_CXL_H define in their build system.