Add fake function to handle sigbus correctly
We like to use that in our maintenance application. In case there is trouble
with the hw, e.g reading registers with -1 values, we do not want to abort
but instead write some error logs before we give up. Without calling that
function libcxl calls abort() (!!!) in those cases and kill the application
using it. Rather drastic I guess.