Trying to get the dependencies more correct to allow parallel build
If we have libcxl.a libcxl.so and the cxl.h dependency in one
line after the all:, I strongly assume that make, when being
called with -j<cpus> is allowed to execute the rules to build that
stuff in parallel. But the cxl.h needs to be there before we can
build the libcxl.so/a or more precisely the libcxl.o object file.

So trying to get this a little more correct than the previous version.
Please have a look.

Signed-off-by: Frank Haverkamp <haver@linux.vnet.ibm.com>