Fix race condition caused by `psl_bfm`
In the current implementation there is a race condition between the call
to `psl_bfm` and all other processes that trigger on `ha_pclock`. Upon
each edge on `ha_pclock` the `psl_bfm` DPI-C function is called, which
read and modifies the `*_top` signals. These changes take effect
immediately in the same delta cycle. However, in this very delta cycle,
other processes run in parallel that read these modified signals. The
SystemVerilog standard does not define what values these processes see,
which is then up to the implementation to decide.

The PSLSE does not work on QuestaSim 10.6, but does work in NCSim (I
don't recall the version). This because QuestaSim handles the resolution
of this race condition in a different manner.

To get around this race condition, have the `psl_bfm` operate on a
temporary copy of each `ha_*_top` signal, and only after the call
returns use a non-blocking signal assignment `ha_*_tmp <= ha_*;` to
schedule the changes for the next delta cycle.